(#5) CommandExecutor: extract tool window management to a separate service
